requirements-vercel.txt
# Vercel-optimized requirements (smaller footprint)
flask==2.3.3
numpy==1.24.3
pillow==10.0.1
scikit-learn==1.3.0
joblib==1.3.2
opencv-python-headless==4.8.1.78

# TensorFlow CPU-only (smaller than full TensorFlow)
tensorflow-cpu==2.15.0